[ETP-2] - Change postfix on test suit from Test to Spec - Change type of volume to BigDecimal - Refactor some class names

+++ src/main/java/com/moebius/entropy/domain/Order.java
@@ -12,5 +12,5 @@ public class Order {
     private final Exchange exchange;
     private final OrderType orderType;
     private final BigDecimal price;
-    private final int volume;
+    private final BigDecimal volume;
 }
